NEXTEUROPA-3527 Add a "dist-clean" target to build the platform.
When building the platform with the  intent to deploy it to a production
environment it  is nice to clean  files that are not  actually needed by
Drupal  at  run  time;  this  includes  test  files,  patches,  scripts,
Git-related  files, examples,  demos,  tools, some  PHP  files and  even
directives for HTTP servers.
The cleaning itself can be refined through the platform.clean property.
Additionally, this dist-clean target takes  care to store the Git commit
into the build.